feat - 수익률 출력 기능

+++ src/main/java/controller/LottoController.java
@@ -74,6 +74,6 @@ public class LottoController {
                 .map(key -> lottoStatistics.get(key) * LottoPrize.valueOf(key))
                 .sum();
 
-        return prizeSum / price;
+        return (double) prizeSum / (double) price;
     }
 }
